Add files via upload
Blackjack is one of the most popular card games in casinos around the world. The great popularity of the game is due to simple rules, the speed of the game and the simplest strategy in card counting.The values of the points of each card: from two to ten - from 2 to 10 respectively, for an ace - 1 or 11, for the so-called. pictures (king, queen, jack) - 10.Thegame lasts until the player scores 21 points in the sum of all cards (Blackjack), or more (then he lost).